fix(core): throw invalid selectors, fix wait x/y
